Fixed bugs, features/accounts part 3
