+++ CryptoBank.WebApi/Features/Accounts/Models/CreateAccountModel.cs
@@ -1,0 +1,7 @@
+﻿namespace CryptoBank.WebApi.Features.Accounts.Models;
+
+public class CreateAccountModel
+{
+    public string Number { get; set; }
+    public string Currency { get; set; }
+}
